Reorganize code into pkg/ directory
