+++ pkg/resource/template/fileStat_posix.go
