+++ pkg/resource/template/fileStat_windows.go
