+++ pkg/resource/template/processor.go
