+++ pkg/resource/template/resource.go
@@ -14,7 +14,7 @@ import (
 	"text/template"
 
 	"github.com/BurntSushi/toml"
-	"github.com/kelseyhightower/confd/backends"
+	"github.com/kelseyhightower/confd/pkg/backends"
 	"github.com/kelseyhightower/memkv"
 	log "github.com/sirupsen/logrus"
 )
