+++ pkg/resource/template/template_funcs.go
