+++ pkg/resource/template/util.go
